trading_calendar_mini/pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "trading-calendar-mini"
version = "0.1.0"
description = "Mini trading calendar: TZ conversions + basic session open/close checks"
requires-python = ">=3.11"
readme = "README.md"
dependencies = ["PyYAML>=6.0"]

[project.scripts]
tzutil = "trading_calendar.cli:tzutil_main"
tcal = "trading_calendar.cli:cal_main"

